remove cyclical deps
