+++ handlers/identity-backfill/src/main/scala/com/gu/identityBackfill/IdentityBackfillSteps.scala
@@ -71,17 +71,3 @@ object IdentityBackfillSteps extends Logging {
 
 }
 
-object Types {
-
-  case class EmailAddress(value: String)
-  case class IdentityId(value: String)
-  case class SFContactId(value: String)
-  case class AccountId(value: String)
-
-  case class ZuoraAccountIdentitySFContact(
-    accountId: AccountId,
-    identityId: Option[IdentityId],
-    sfContactId: SFContactId
-  )
-
-}
